internal/adapters/rest/handlers/parking_lot_delete_test.go
package handlers_test

import (
	"bytes"
	"context"
	"net/http"
	"net/http/httptest"
	"testing"

	"github.com/IgorSteps/easypark/internal/adapters/rest/handlers"
	"github.com/IgorSteps/easypark/internal/domain/repositories"
	mocks "github.com/IgorSteps/easypark/mocks/adapters/rest/handlers"
	"github.com/go-chi/chi/v5"
	"github.com/google/uuid"
	"github.com/sirupsen/logrus/hooks/test"
	"github.com/stretchr/testify/assert"
)

func TestDeleteParkingLotHandler_ServeHTTP_HappyPath(t *testing.T) {
	// --------
	// ASSEMBLE
	// --------
	testLogger, _ := test.NewNullLogger()
	mockFacade := &mocks.ParkingLotFacade{}
	handler := handlers.NewDeleteParkingLotHandler(testLogger, mockFacade)

	testID := uuid.New()

	// Because we are directly calling the handler in the test without going through a router that parses the URL parameters,
	// we have to manually insert the URL parameters into the request context.
	rctx := chi.NewRouteContext()
	rctx.URLParams.Add("id", testID.String())

	// Create our request context with the formatted chi context.
	reqCtx := context.WithValue(context.Background(), chi.RouteCtxKey, rctx)
	req, _ := http.NewRequestWithContext(reqCtx, "DELETE", "/parking-logs", bytes.NewBuffer(nil))
	rr := httptest.NewRecorder()

	mockFacade.EXPECT().DeleteParkingLot(req.Context(), testID).Return(nil).Once()

	// --------
	// ACT
	// --------
	handler.ServeHTTP(rr, req)

	// --------
	// ASSERT
	// --------
	assert.Equal(t, http.StatusOK, rr.Code, "Response codes don't match, should be 200")
	assert.Contains(t, rr.Body.String(), "successfully deleted parking lot", "Reponse bodies don't match")
	mockFacade.AssertExpectations(t)
}

func TestDeleteParkingLotHandler_ServeHTTP_UnhappyPath_InvalidInputError(t *testing.T) {
	// --------
	// ASSEMBLE
	// --------
	testLogger, _ := test.NewNullLogger()
	mockFacade := &mocks.ParkingLotFacade{}
	handler := handlers.NewDeleteParkingLotHandler(testLogger, mockFacade)

	testID := uuid.New()

	// Because we are directly calling the handler in the test without going through a router that parses the URL parameters,
	// we have to manually insert the URL parameters into the request context.
	rctx := chi.NewRouteContext()
	rctx.URLParams.Add("id", testID.String())

	// Create our request context with the formatted chi context.
	reqCtx := context.WithValue(context.Background(), chi.RouteCtxKey, rctx)
	req, _ := http.NewRequestWithContext(reqCtx, "DELETE", "/parking-logs", bytes.NewBuffer(nil))
	rr := httptest.NewRecorder()

	testError := repositories.NewInvalidInputError("boom")
	mockFacade.EXPECT().DeleteParkingLot(req.Context(), testID).Return(testError).Once()

	// --------
	// ACT
	// --------
	handler.ServeHTTP(rr, req)

	// --------
	// ASSERT
	// --------
	assert.Equal(t, http.StatusBadRequest, rr.Code, "Response codes don't match, should be 400")
	assert.Contains(t, rr.Body.String(), "boom", "Reponse bodies don't match")
	mockFacade.AssertExpectations(t)
}

func TestDeleteParkingLotHandler_ServeHTTP_UnhappyPath_InternalError(t *testing.T) {
	// --------
	// ASSEMBLE
	// --------
	testLogger, _ := test.NewNullLogger()
	mockFacade := &mocks.ParkingLotFacade{}
	handler := handlers.NewDeleteParkingLotHandler(testLogger, mockFacade)

	testID := uuid.New()

	// Because we are directly calling the handler in the test without going through a router that parses the URL parameters,
	// we have to manually insert the URL parameters into the request context.
	rctx := chi.NewRouteContext()
	rctx.URLParams.Add("id", testID.String())

	// Create our request context with the formatted chi context.
	reqCtx := context.WithValue(context.Background(), chi.RouteCtxKey, rctx)
	req, _ := http.NewRequestWithContext(reqCtx, "DELETE", "/parking-logs", bytes.NewBuffer(nil))
	rr := httptest.NewRecorder()

	testError := repositories.NewInternalError("boom")
	mockFacade.EXPECT().DeleteParkingLot(req.Context(), testID).Return(testError).Once()

	// --------
	// ACT
	// --------
	handler.ServeHTTP(rr, req)

	// --------
	// ASSERT
	// --------
	assert.Equal(t, http.StatusInternalServerError, rr.Code, "Response codes don't match, should be 500")
	assert.Contains(t, rr.Body.String(), "boom", "Reponse bodies don't match")
	mockFacade.AssertExpectations(t)
}
